test,win: divide the addon modules test into two
The original addons/esm test was failing when building the addon after
updating V8 to v13.4 on Windows. After an investigation, it was clear
that the targets of the binding.gyp cannot coexist together so it was
divided into 2 sets of 2 targets with no overlapping.
