+++ test/addons/esm-export/binding-export-default.cc
