+++ test/addons/esm-export/binding-export-primitive.cc
