+++ test/addons/esm/binding.gyp
@@ -1,21 +1,11 @@
 {
   'variables': {
-    'source_dir': '<!(<(python) -c "import os; print(os.getcwd())")',
+    'source_dir': '<!("<(python)" -c "import os; print(os.getcwd())")',
   },
   'targets': [
     {
       'target_name': 'binding',
       'sources': [ 'binding.cc' ],
-      'includes': ['../common.gypi'],
-    },
-    {
-      'target_name': 'binding-export-default',
-      'sources': [ 'binding-export-default.cc' ],
-      'includes': ['../common.gypi'],
-    },
-    {
-      'target_name': 'binding-export-primitive',
-      'sources': [ 'binding-export-primitive.cc' ],
       'includes': ['../common.gypi'],
     },
     {
