+++ test/addons/esm/test-esm.mjs
@@ -31,27 +31,4 @@ export async function run() {
     assert.strictEqual(rebinding.hello(), 'world');
     assert.strictEqual(binding.hello, rebinding.hello);
   }
-
-  // binding-export-default.node
-  {
-    const bindingPath = require.resolve(`./build/${buildType}/binding-export-default.node`);
-    // Test with order of require+import
-    const bindingRequire = require(bindingPath);
-    const ns = await import(pathToFileURL(bindingPath));
-    assert.strictEqual(ns.default, bindingRequire);
-
-    // As same as ESM-import-CJS, the default export is the value of `module.exports`.
-    assert.strictEqual(ns.default, ns['module.exports']);
-    assert.strictEqual(ns.default.default(), 'hello world');
-  }
-
-  // binding-export-primitive.node
-  {
-    const bindingPath = require.resolve(`./build/${buildType}/binding-export-primitive.node`);
-    const ns = await import(pathToFileURL(bindingPath));
-
-    // As same as ESM-import-CJS, the default export is the value of `module.exports`.
-    assert.strictEqual(ns.default, ns['module.exports']);
-    assert.strictEqual(ns.default, 'hello world');
-  }
 }
